requirements.txt
coreapi==2.3.0
coreschema==0.0.4
Django==1.11
djangorestframework==3.6.2
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
Pygments==2.2.0
pytz==2017.2
requests==2.13.0
uritemplate==3.0.0
